+++ index.html
@@ -4,9 +4,9 @@
  <meta charset="UTF-8">
  <title>Weather.IM Live</title>
  <link rel="icon" type="image/x-icon" href="favicon.ico">
- <link rel="stylesheet" type="text/css" href="vendor/extjs-gpl/build/classic/theme-neptune/resources/theme-neptune-all.css"/>
+ <link rel="stylesheet" type="text/css" href="vendor/avalos-extjs-gpl/build/classic/theme-neptune/resources/theme-neptune-all.css"/>
  <link rel="stylesheet" type="text/css" href="src/live.css"/>
- <script type="text/javascript" src="vendor/extjs-gpl/build/ext-all.js"></script>
+ <script type="text/javascript" src="vendor/avalos-extjs-gpl/build/ext-all.js"></script>
 </head>
 <body>
 
@@ -28,10 +28,15 @@
 <a target="_blank" href="https://github.com/akrherz/live">akrherz/live</a>
 Filing issues, submitting pull requests is certainly welcome.  You should be
 able to run this app locally via <code>npm start</code> after cloning the
-repo.</p>
+repo.  This app is licensed under the GPLv3 to comply with the required
+license when using ExtJS-GPL.</p>
 
 <p><h4>Most Recent Changes</h4>
 <ul>
+<li>
+    <strong>6 Mar 2026</strong>: Updated to ExtJS 7.0 to fix CSS troubles
+    <a target="_blank" href="https://github.com/akrherz/live/issues/82">GH82</a>
+</li>
 <li>
     <strong>5 Mar 2026</strong>: Fix auto-scrolling to latest message (hopefully)
     <a target="_blank" href="https://github.com/akrherz/live/issues/77">GH77</a>
@@ -44,18 +49,6 @@ repo.</p>
     <strong>28 Feb 2026</strong>: Restore no map option when called with ?nomap
     <a target="_blank" href="https://github.com/akrherz/live/issues/73">GH73</a>
 </li>
-<li>
-    <strong>28 Feb 2026</strong>: Fix chatroom bookmarks.
-    <a target="_blank" href="https://github.com/akrherz/live/issues/72">GH72</a>
-</li>
-<li>
-    <strong>27 Feb 2026</strong>: Restored political boundary layers.
-    <a target="_blank" href="https://github.com/akrherz/live/issues/69">GH69</a>
-</li>
-    <li><strong>27 Feb 2026</strong>: 12 years later, a new version with
- many new bugs, likely.</li>
- <li>12 Jun 2014: Fix bug where warnings and local storm reports were not
- clickable on the map.</li>
 </ul>
 
 <p style="margin-top: 5px;">"Weather.IM Live" is a pure web browser
