+++ package.json
@@ -10,7 +10,7 @@
     "type": "git",
     "url": "git+https://github.com/akrherz/live.git"
   },
-  "license": "ISC",
+  "license": "GPL-3.0-only",
   "author": "",
   "type": "module",
   "scripts": {
@@ -18,10 +18,6 @@
     "test:coverage": "mkdir -p coverage && node --test --experimental-test-coverage --test-reporter=spec --test-reporter-destination=stderr --test-reporter=lcov --test-reporter-destination=coverage/lcov.info tests/layer-config.test.mjs tests/date-utils.test.mjs",
     "test:layer-config": "node --test tests/layer-config.test.mjs",
     "test:date-utils": "node --test tests/date-utils.test.mjs",
-    "sync:extjs": "node scripts/sync-extjs-assets.mjs",
-    "prestart": "npm run sync:extjs",
-    "predev": "npm run sync:extjs",
-    "prebuild": "npm run sync:extjs",
     "start": "vite",
     "dev": "vite",
     "build": "vite build",
@@ -31,7 +27,7 @@
   },
   "dependencies": {
     "@eslint/js": "^9.39.2",
-    "extjs-gpl": "^6.2.0",
+    "@avalos/extjs-gpl": "^7.0.0",
     "globals": "^17.0.0",
     "iemjs": "^0.0.7",
     "ol": "^10.7.0",
